reorganize-files: rename App to MemeCreator, place in a folder
