+++ src/01_MemeCreator/MemeCreator.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import { Image, CanvasRenderingContext2D } from 'canvas';
+import { Image } from 'canvas';
 import { fireEvent, render, act } from '@testing-library/react';
 import { saveAs } from 'file-saver';
 
-import App from './AppHooks';
+import MemeCreator from './MemeCreator';
 
 jest.mock('./memeTemplates.json', () => [
   {
@@ -36,7 +36,7 @@ test('renders component correctly', async () => {
   let getByLabelText;
 
   await act(async () => {
-    ({ getByLabelText } = render(<App />));
+    ({ getByLabelText } = render(<MemeCreator />));
   });
 
   const selectElement = getByLabelText(/meme template/i);
@@ -47,17 +47,14 @@ test('renders component correctly', async () => {
 
 test('sets up the canvas on load', async () => {
   let container;
-  const drawImgSpy = jest.spyOn(CanvasRenderingContext2D.prototype, 'drawImage');
-  const fillTxtSpy = jest.spyOn(CanvasRenderingContext2D.prototype, 'fillText');
+  const drawSpy = jest.spyOn(MemeCreator.prototype, 'drawCanvas');
 
   await act(async () => {
-    ({ container } = render(<App />));
+    ({ container } = render(<MemeCreator />));
   });
 
-  expect(drawImgSpy).toHaveBeenCalledTimes(1);
-  expect(drawImgSpy).toHaveBeenCalledWith(expect.any(Image), 0, 0);
-  expect(fillTxtSpy).toHaveBeenCalledTimes(1);
-  expect(fillTxtSpy).toHaveBeenCalledWith('', expect.any(Number), expect.any(Number));
+  expect(drawSpy).toHaveBeenCalledTimes(1);
+  expect(drawSpy).toHaveBeenCalledWith(expect.any(Image), '');
 
   const canvas = container.querySelector('canvas');
   expect(canvas.width).toBe(500); // match dawson.jpg width
@@ -65,21 +62,20 @@ test('sets up the canvas on load', async () => {
   const dataUrl = canvas.toDataURL('image/png', 0.92);
   expect(dataUrl).toMatchSnapshot();
 
-  drawImgSpy.mockRestore();
-  fillTxtSpy.mockRestore();
+  drawSpy.mockRestore();
 });
 
 test('updates the canvas on caption change', async () => {
   let container;
   let getByLabelText;
-  const fillTxtSpy = jest.spyOn(CanvasRenderingContext2D.prototype, 'fillText');
+  const drawSpy = jest.spyOn(MemeCreator.prototype, 'drawCanvas');
 
   await act(async () => {
-    ({ container, getByLabelText } = render(<App />));
+    ({ container, getByLabelText } = render(<MemeCreator />));
   });
 
-  expect(fillTxtSpy).toHaveBeenCalledTimes(1);
-  expect(fillTxtSpy).toHaveBeenCalledWith('', expect.any(Number), expect.any(Number));
+  expect(drawSpy).toHaveBeenCalledTimes(1);
+  expect(drawSpy).toHaveBeenLastCalledWith(expect.any(Image), '');
 
   // update caption
   await act(async () => {
@@ -87,8 +83,8 @@ test('updates the canvas on caption change', async () => {
     fireEvent.change(inputElement, { target: { value: 'Hello' } });
   });
 
-  expect(fillTxtSpy).toHaveBeenCalledTimes(2);
-  expect(fillTxtSpy).toHaveBeenLastCalledWith('Hello', expect.any(Number), expect.any(Number));
+  expect(drawSpy).toHaveBeenCalledTimes(2);
+  expect(drawSpy).toHaveBeenLastCalledWith(expect.any(Image), 'Hello');
 
   const canvas = container.querySelector('canvas');
   expect(canvas.width).toBe(500); // match dawson.jpg width
@@ -96,20 +92,20 @@ test('updates the canvas on caption change', async () => {
   const dataUrl = canvas.toDataURL('image/png', 0.92);
   expect(dataUrl).toMatchSnapshot();
 
-  fillTxtSpy.mockRestore();
+  drawSpy.mockRestore();
 });
 
 test('updates the canvas when another template is selected', async () => {
   let container;
   let getByLabelText;
-  const drawImgSpy = jest.spyOn(CanvasRenderingContext2D.prototype, 'drawImage');
+  const drawSpy = jest.spyOn(MemeCreator.prototype, 'drawCanvas');
 
   await act(async () => {
-    ({ container, getByLabelText } = render(<App />));
+    ({ container, getByLabelText } = render(<MemeCreator />));
   });
 
-  expect(drawImgSpy).toHaveBeenCalledTimes(1);
-  expect(drawImgSpy).toHaveBeenCalledWith(expect.any(Image), 0, 0);
+  expect(drawSpy).toHaveBeenCalledTimes(1);
+  expect(drawSpy).toHaveBeenLastCalledWith(expect.any(Image), '');
 
   // update template
   await act(async () => {
@@ -117,8 +113,8 @@ test('updates the canvas when another template is selected', async () => {
     fireEvent.change(selectElement, { target: { value: 'jackieChan' } });
   });
 
-  expect(drawImgSpy).toHaveBeenCalledTimes(2);
-  expect(drawImgSpy).toHaveBeenLastCalledWith(expect.any(Image), 0, 0);
+  expect(drawSpy).toHaveBeenCalledTimes(2);
+  expect(drawSpy).toHaveBeenLastCalledWith(expect.any(Image), '');
 
   const canvas = container.querySelector('canvas');
   expect(canvas.width).toBe(500); // match jackie-chan.jpg width
@@ -126,7 +122,7 @@ test('updates the canvas when another template is selected', async () => {
   const dataUrl = canvas.toDataURL('image/png', 0.92);
   expect(dataUrl).toMatchSnapshot();
 
-  drawImgSpy.mockRestore();
+  drawSpy.mockRestore();
 });
 
 test('triggers a save of the canvas Blob when clicking Download btn', async () => {
@@ -138,7 +134,7 @@ test('triggers a save of the canvas Blob when clicking Download btn', async () =
 
   let getByText;
   await act(async () => {
-    ({ getByText } = render(<App />));
+    ({ getByText } = render(<MemeCreator />));
   });
 
   // click Download btn
