+++ src/01_MemeCreator/memeTemplates.json
